+++ _wasm_crypto/mod.ts
@@ -26,12 +26,14 @@ export const digestAlgorithms = [
   "SHA3-512",
   "SHAKE128",
   "SHAKE256",
+  "TIGER",
   // insecure (length-extendable):
   "RIPEMD-160",
   "SHA-224",
   "SHA-256",
   "SHA-512",
   // insecure (collidable and length-extendable):
+  "MD4",
   "MD5",
   "SHA-1",
 ] as const;