+++ log/handlers.ts
@@ -102,9 +102,9 @@ export class FileHandler extends WriterHandler {
   protected _mode: LogMode;
   protected _openOptions: Deno.OpenOptions;
   protected _encoder = new TextEncoder();
-  #unloadCallback() {
+  #unloadCallback = (() => {
     this.destroy();
-  }
+  }).bind(this);
 
   constructor(levelName: LevelName, options: FileHandlerOptions) {
     super(levelName, options);
@@ -125,7 +125,7 @@ export class FileHandler extends WriterHandler {
     this._writer = this._file;
     this._buf = new BufWriterSync(this._file);
 
-    addEventListener("unload", this.#unloadCallback.bind(this));
+    addEventListener("unload", this.#unloadCallback);
   }
 
   handle(logRecord: LogRecord): void {
@@ -138,6 +138,9 @@ export class FileHandler extends WriterHandler {
   }
 
   log(msg: string): void {
+    if (this._encoder.encode(msg).byteLength + 1 > this._buf.available()) {
+      this.flush();
+    }
     this._buf.writeSync(this._encoder.encode(msg + "\n"));
   }
 
@@ -214,7 +217,8 @@ export class RotatingFileHandler extends FileHandler {
       this.#currentFileSize = 0;
     }
 
-    this._buf.writeSync(this._encoder.encode(msg + "\n"));
+    super.log(msg);
+
     this.#currentFileSize += msgByteLength;
   }