+++ node/_readline.js
@@ -29,8 +29,8 @@ import {
 } from "./internal/readline/callbacks.js";
 import { emitKeypressEvents } from "./internal/readline/emitKeypressEvents.js";
 import { validateAbortSignal } from "./internal/validators.js";
-import { promisify } from "./_util/_util_promisify.ts";
-import { AbortError } from "./_errors.ts";
+import { promisify } from "./internal/util.js";
+import { AbortError } from "./internal/errors.ts";
 
 import {
   Interface as _Interface,