+++ node/fs.ts
@@ -33,9 +33,12 @@ import { truncate, truncateSync } from "./_fs/_fs_truncate.ts";
 import { unlink, unlinkSync } from "./_fs/_fs_unlink.ts";
 import { utimes, utimesSync } from "./_fs/_fs_utimes.ts";
 import { watch, watchFile } from "./_fs/_fs_watch.ts";
+// @deno-types="./_fs/_fs_write.d.ts"
+import { write, writeSync } from "./_fs/_fs_write.js";
 import { writeFile, writeFileSync } from "./_fs/_fs_writeFile.ts";
+import { Stats } from "./internal/fs/utils.js";
 
-import * as promises from "./fs/promises.ts";
+import { promisify } from "./util.ts";
 
 const {
   F_OK,
@@ -43,6 +46,37 @@ const {
   W_OK,
   X_OK,
 } = constants;
+
+const promises = {
+  access: promisify(access),
+  copyFile: promisify(copyFile),
+  open: promisify(open),
+  // opendir: promisify(opendir),
+  rename: promisify(rename),
+  truncate: promisify(truncate),
+  rm: promisify(rm),
+  rmdir: promisify(rmdir),
+  mkdir: promisify(mkdir),
+  readdir: promisify(readdir),
+  readlink: promisify(readlink),
+  symlink: promisify(symlink),
+  lstat: promisify(lstat),
+  stat: promisify(stat),
+  link: promisify(link),
+  unlink: promisify(unlink),
+  chmod: promisify(chmod),
+  // lchmod: promisify(lchmod),
+  // lchown: promisify(lchown),
+  chown: promisify(chown),
+  utimes: promisify(utimes),
+  // lutimes = promisify(lutimes),
+  realpath: promisify(realpath),
+  mkdtemp: promisify(mkdtemp),
+  writeFile: promisify(writeFile),
+  appendFile: promisify(appendFile),
+  readFile: promisify(readFile),
+  watch: promisify(watch),
+};
 
 export default {
   access,
@@ -102,6 +136,7 @@ export default {
   rm,
   rmSync,
   stat,
+  Stats,
   statSync,
   symlink,
   symlinkSync,
@@ -114,6 +149,8 @@ export default {
   W_OK,
   watch,
   watchFile,
+  write,
+  writeSync,
   writeFile,
   writeFileSync,
   X_OK,
@@ -177,6 +214,7 @@ export {
   rmdirSync,
   rmSync,
   stat,
+  Stats,
   statSync,
   symlink,
   symlinkSync,
@@ -189,7 +227,9 @@ export {
   W_OK,
   watch,
   watchFile,
+  write,
   writeFile,
   writeFileSync,
+  writeSync,
   X_OK,
 };