+++ node/http.ts
@@ -1,7 +1,7 @@
 import { core } from "./_core.ts";
 import { _normalizeArgs, ListenOptions } from "./net.ts";
 import { Buffer } from "./buffer.ts";
-import { ERR_SERVER_NOT_RUNNING } from "./_errors.ts";
+import { ERR_SERVER_NOT_RUNNING } from "./internal/errors.ts";
 import { EventEmitter } from "./events.ts";
 import { nextTick } from "./_next_tick.ts";
 import { Status as STATUS_CODES } from "../http/http_status.ts";
@@ -10,6 +10,9 @@ import {
   Readable as NodeReadable,
   Writable as NodeWritable,
 } from "./stream.ts";
+import { OutgoingMessage } from "./_http_outgoing.ts";
+import { Agent } from "./_http_agent.js";
+import { urlToHttpOptions } from "./internal/url.ts";
 
 const METHODS = [
   "ACL",
@@ -55,6 +58,127 @@ function chunkToU8(chunk: Chunk): Uint8Array {
     return core.encode(chunk);
   }
   return chunk;
+}
+
+export interface RequestOptions {
+  agent?: Agent;
+  auth?: string;
+  createConnection?: () => unknown;
+  defaultPort?: number;
+  family?: number;
+  headers?: Record<string, string>;
+  hints?: number;
+  host?: string;
+  hostname?: string;
+  insecureHTTPParser?: boolean;
+  localAddress?: string;
+  localPort?: number;
+  lookup?: () => void;
+  maxHeaderSize?: number;
+  method?: string;
+  path?: string;
+  port?: number;
+  protocol?: string;
+  setHost?: boolean;
+  socketPath?: string;
+  timeout?: number;
+  signal?: AbortSignal;
+  href?: string;
+}
+
+/** ClientRequest represents the http(s) request from the client */
+class ClientRequest extends NodeWritable {
+  body: null | ReadableStream = null;
+  controller: ReadableStreamDefaultController | null = null;
+  constructor(
+    public opts: RequestOptions,
+    public cb: (res: IncomingMessageForClient) => void,
+  ) {
+    super();
+  }
+
+  // deno-lint-ignore no-explicit-any
+  _write(chunk: any, _enc: string, cb: () => void) {
+    if (this.controller) {
+      this.controller.enqueue(chunk);
+      cb();
+      return;
+    }
+
+    this.body = new ReadableStream({
+      start: (controller) => {
+        this.controller = controller;
+        controller.enqueue(chunk);
+        cb();
+      },
+    });
+  }
+
+  async _final() {
+    const client = await this._createCustomClient();
+    const opts = { body: this.body, method: this.opts.method, client };
+    const res = new IncomingMessageForClient(
+      await fetch(this.opts.href!, opts),
+    );
+    this.emit("response", res);
+    if (client) {
+      res.on("end", () => {
+        client.close();
+      });
+    }
+    this.cb(res);
+  }
+
+  abort() {
+    this.destroy();
+  }
+
+  _createCustomClient(): Promise<Deno.HttpClient | undefined> {
+    return Promise.resolve(undefined);
+  }
+}
+
+/** IncomingMessage for http(s) client */
+export class IncomingMessageForClient extends NodeReadable {
+  reader: ReadableStreamDefaultReader | undefined;
+  constructor(public resp: Response) {
+    super();
+    this.reader = resp.body?.getReader();
+  }
+
+  async _read(_size: number) {
+    if (this.reader === undefined) {
+      this.push(null);
+      return;
+    }
+    try {
+      const res = await this.reader.read();
+      if (res.done) {
+        this.push(null);
+        return;
+      }
+      this.push(res.value);
+    } catch (e) {
+      // deno-lint-ignore no-explicit-any
+      this.destroy(e as any);
+    }
+  }
+
+  get headers() {
+    return Object.fromEntries(this.resp.headers.entries());
+  }
+
+  get trailers() {
+    return {};
+  }
+
+  get statusCode() {
+    return this.resp.status;
+  }
+
+  get statusMessage() {
+    return this.resp.statusText;
+  }
 }
 
 export class ServerResponse extends NodeWritable {
@@ -162,7 +286,7 @@ export class ServerResponse extends NodeWritable {
 
   // deno-lint-ignore no-explicit-any
   end(chunk?: any, encoding?: any, cb?: any): this {
-    if (!chunk) {
+    if (!chunk && this.#headers.has("transfer-encoding")) {
       // FIXME(bnoordhuis) Node sends a zero length chunked body instead, i.e.,
       // the trailing "0\r\n", but respondWith() just hangs when I try that.
       this.#headers.set("content-length", "0");
@@ -175,7 +299,7 @@ export class ServerResponse extends NodeWritable {
 }
 
 // TODO(@AaronO): optimize
-export class IncomingMessage extends NodeReadable {
+export class IncomingMessageForServer extends NodeReadable {
   private req: Request;
   url: string;
 
@@ -223,9 +347,16 @@ export class IncomingMessage extends NodeReadable {
   }
 }
 
-type ServerHandler = (req: IncomingMessage, res: ServerResponse) => void;
+type ServerHandler = (
+  req: IncomingMessageForServer,
+  res: ServerResponse,
+) => void;
 
-export class Server extends EventEmitter {
+export function Server(handler?: ServerHandler): ServerImpl {
+  return new ServerImpl(handler);
+}
+
+class ServerImpl extends EventEmitter {
   #httpConnections: Set<Deno.HttpConn> = new Set();
   #listener?: Deno.Listener;
 
@@ -282,7 +413,7 @@ export class Server extends EventEmitter {
           if (reqEvent === null) {
             break;
           }
-          const req = new IncomingMessage(reqEvent.request);
+          const req = new IncomingMessageForServer(reqEvent.request);
           const res = new ServerResponse(reqEvent);
           this.emit("request", req, res);
         }
@@ -356,16 +487,73 @@ export class Server extends EventEmitter {
   }
 }
 
+Server.prototype = ServerImpl.prototype;
+
 export function createServer(handler?: ServerHandler) {
-  return new Server(handler);
+  return Server(handler);
 }
 
-export { METHODS, STATUS_CODES };
+/** Makes an HTTP request. */
+export function request(
+  url: string | URL,
+  cb?: (res: IncomingMessageForClient) => void,
+): ClientRequest;
+export function request(
+  opts: RequestOptions,
+  cb?: (res: IncomingMessageForClient) => void,
+): ClientRequest;
+export function request(
+  url: string | URL,
+  opts: RequestOptions,
+  cb?: (res: IncomingMessageForClient) => void,
+): ClientRequest;
+// deno-lint-ignore no-explicit-any
+export function request(...args: any[]) {
+  let options = {};
+  if (typeof args[0] === "string") {
+    options = urlToHttpOptions(new URL(args.shift()));
+  } else if (args[0] instanceof URL) {
+    options = urlToHttpOptions(args.shift());
+  }
+  if (args[0] && typeof args[0] !== "function") {
+    Object.assign(options, args.shift());
+  }
+  args.unshift(options);
+  return new ClientRequest(args[0], args[1]);
+}
+
+/** Makes a `GET` HTTP request. */
+export function get(
+  url: string | URL,
+  cb?: (res: IncomingMessageForClient) => void,
+): ClientRequest;
+export function get(
+  opts: RequestOptions,
+  cb?: (res: IncomingMessageForClient) => void,
+): ClientRequest;
+export function get(
+  url: string | URL,
+  opts: RequestOptions,
+  cb?: (res: IncomingMessageForClient) => void,
+): ClientRequest;
+// deno-lint-ignore no-explicit-any
+export function get(...args: any[]) {
+  const req = request(args[0], args[1], args[2]);
+  req.end();
+  return req;
+}
+
+export { Agent, ClientRequest, METHODS, OutgoingMessage, STATUS_CODES };
 export default {
+  Agent,
+  ClientRequest,
   STATUS_CODES,
   METHODS,
   createServer,
   Server,
-  IncomingMessage,
+  IncomingMessage: IncomingMessageForServer,
+  OutgoingMessage,
   ServerResponse,
+  request,
+  get,
 };