+++ node/vm.ts
@@ -5,12 +5,13 @@
 import { notImplemented } from "./_utils.ts";
 
 export class Script {
+  code: string;
   constructor(code: string, _options = {}) {
-    code = `${code}`;
+    this.code = `${code}`;
   }
 
   runInThisContext(_options: any) {
-    notImplemented();
+    return eval.call(globalThis, this.code);
   }
 
   runInContext(_contextifiedObject: any, _options: any) {
@@ -51,10 +52,10 @@ export function runInNewContext(
 }
 
 export function runInThisContext(
-  _code: string,
-  _options: any,
+  code: string,
+  options: any,
 ) {
-  notImplemented();
+  return createScript(code, options).runInThisContext(options);
 }
 
 export function isContext(_maybeContext: any) {