+++ uuid/mod.ts
@@ -23,4 +23,39 @@ export function isNil(id: string): boolean {
   return id === NIL_UUID;
 }
 
+/**
+ * Test a string to see if it is a valid UUID.
+ *
+ * ```js
+ * import { validate } from "./mod.ts"
+ *
+ * validate("not a UUID") // false
+ * validate("6ec0bd7f-11c0-43da-975e-2a8ad9ebae0b") // true
+ * ```
+ */
+export function validate(uuid: string): boolean {
+  return /^(?:[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}|00000000-0000-0000-0000-000000000000)$/i
+    .test(
+      uuid,
+    );
+}
+
+/**
+ * Detect RFC version of a UUID.
+ *
+ * ```js
+ * import { version } from "./mod.ts"
+ *
+ * version("d9428888-122b-11e1-b85c-61cd3cbb3210") // 1
+ * version("109156be-c4fb-41ea-b1b4-efe1671c5836") // 4
+ * ```
+ */
+export function version(uuid: string): number {
+  if (!validate(uuid)) {
+    throw TypeError("Invalid UUID");
+  }
+
+  return parseInt(uuid[14], 16);
+}
+
 export { v1, v4, v5 };